feat: better theta grid as specified by Jochen
